+++ src/pages/oracle/landing/components/TitleText/TitleText.module.scss
