+++ src/pages/oracle/landing/components/TitleText/TitleText.tsx
