+++ src/pages/oracle/landing/graphDataPrepared.json
